refactor(store): replace root barrel imports with owning modules
Move all 16 remaining root store consumers to their canonical atom and type definitions, remove the unused root store barrel, and update import guidance. Atom implementations and UI behavior are unchanged. Verified typechecking, 23 targeted tests, circular dependencies and stale module references.
