+++ src/hooks/navigation/useGlobalShortcuts/useTabShortcuts.ts
@@ -14,7 +14,6 @@ import {
 } from "@src/scaffold/GlobalSpotlight/openSpotlight";
 import { AppViewService } from "@src/services/app";
 import { WorkStationViewService } from "@src/services/workStation/WorkStationViewService";
-import { spotlightInitialQueryAtom, spotlightOpenAtom } from "@src/store";
 import { openOrFocusChatPanelStartPageTabAtom } from "@src/store/chatPanel/chatPanelTabsAtom";
 import { modelSelectorAtom } from "@src/store/ui/modelSelectorAtom";
 import {
@@ -23,6 +22,10 @@ import {
   repoSelectorOpenAtom,
 } from "@src/store/ui/overlayAtom";
 import { sidebarCollapsedAtom } from "@src/store/ui/sidebarAtom";
+import {
+  spotlightInitialQueryAtom,
+  spotlightOpenAtom,
+} from "@src/store/ui/uiAtom";
 import { closeActiveWorkStationTabAtom } from "@src/store/workstation/tabRegistry";
 
 /**
