+++ src/scaffold/ModalSystem/variants/RouteDebug/index.tsx
@@ -3,8 +3,8 @@ import React, { useEffect } from "react";
 import { useLocation, useMatches, useParams } from "react-router-dom";
 
 import Message from "@src/components/Message";
-import { routeDebugModalOpenAtom } from "@src/store";
 import { devModeEnabledAtom } from "@src/store/platform/devModeAtom";
+import { routeDebugModalOpenAtom } from "@src/store/ui/uiAtom";
 import { getInstrumentedStore } from "@src/util/core/state/instrumentedStore";
 
 function buildRouteText(
